+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Fix `UserData.mutedKeywordsRegex` when the user has no server side filters
 
 ### v0.72.2
 * Add `lastEditedAt` getter to `Toot` and compare it to `completedAt` when evaluating `Toot` completeness so as to catch edits